persistence-service/requirements.txt
confluent-kafka
python-dotenv
SQLAlchemy
psycopg2-binary